requirements.txt
openai
python-docx
python-dotenv
pywin32
gradio
